Drawing now works better + input setup
The drawing code makes more sense and allows for new ways to draw to be implemented quickly. The Enter button, used to begin forward propagation of the network, now turns the 'drawing' into a list of pixels with brightness values ranging from 0-255, matching the MNIST data set.
The next step is to use NumPy to define the neural network and write mathematical functions which generate a new network with random initial values, feed forward and backpropagate ("learning"). Once this  is completed, I will add the ability to save and load a neural network model containing predefined weights and biases. This will be probably be stored in the folder of the python project using numpy saving functions for simplicity.