Feed forward with softmax and ReLU
This is really exciting if it works, which I am not yet sure of.
I wrote code for a network class which can hold a neural network and its functions.

Inside this class, I wrote
code that creates sets of weights and biases for the amount of nuerons in each of the layers present in the network.

Then, I wrote a function which iterates through the sets of weights and biases when given the input layer. Initially, the function takes the input layer matrix. Then, it calculates the weighted sum accounting for the biases of each value in the matrix. Since the code is in a for loop, the input layer variable is overwritten with this result. This process is then repeated with the next set of weights and biases until the output layer is reached (with last set of weights and biases). Then, the for loop is finished, and a softmax function can be applied. The softmax function is good for using on the output layer as opposed to a ReLU because it turns the values of the ten outputs into a probability matrix where the highest number output has the highest percent probability. (Note that after the last weighted sum is calculated, the loop must apply the softmax and NOT the ReLU. Applying a softmax on top of a ReLU is, according to a random programmer on Stack Overflow, not logical. This is a good way of showing that the network is guessing the digit with a certain confidence and the other digits it thinks it less closely matches.