Network feed forward code WORKS!!!!
The network will now actually produce output values when given an input drawing.

The input drawing is converted into normalized grayscale values for the network. The network feeds forward and the output layer produces probabilities based on the magnitude of each activation using a softmax function.

It is so cool to see that this finally works. It took a lot of debugging and Googling since I have never used NumPy before.

One noteable change besides the debugging is that I changed the network to initialize the weights and biases using values from the intervals [-0.1, 0.1] instead of [0, 1), which is the default in numpy. This is because, when using no negative values and a linear activation function such as the ReLU, the neuron values become large enough to exceed the 64 bits. In the future I may change the numpy arrays to use 128-bit floats, but I will only do that if it is proven necessary when I train a model using the MNIST dataset.

I now need to finish implementing backpropagation. I converted the mathematical formulas for the relevant derivatives into python code, and I now need to create a function which will iterate through the MNIST database's set of handwritten digit images and adjust the network's parameters to make the network consistently guess the digit being drawn.

Some other things TO DO at this time:

Use numpy to create functions to save a trained network model to a file and load them, so re-training the network each time the program is quit and reopened is not necessary.

Allow the user of the program to use the GUI to say if the network incorrectly guessed the digit they drew. If the guess was incorrect, they can input the correct digit, and the network will backpropagate using the drawing and label as an example, improving the accuracy of the network after receiving real-world feedback.