added MATPLOTLIB ANIMATIONS!
this is sooooo cool! I will work on adding animations in a more complex way when I turn the code into a model-object based system. Currently there is no model object, only layer, activation, loss, and optimizer objects which you have to manually tell to compute the forward and backward pass and run calculations. Once the model object format is used, it will be much more like pytorch where you build the network model by listing the layers and stuff and then then you can tell it to train on a dataset or compute a forward pass on data with a few lines of code. This also will introduce the functionality of saving and loading network parameters.