Made app for clothing item drawing classification.
I know this is not handwritten digits, but it's more interesting.

Results: Based on having my family try it it seems like the preprocessing algorithm needs to be improved. The network was trained on images which are scaled so the content is perfectly entered and takes up the whole image.

To Improve: I am going to use OpenCV contours to find the positive space portion of the drawing and scale it to fit the entire image. This should improve results dramatically and make it much more intuitive for the user.