Greatly improved preprocessing with new step!
Added an additional step to the preprocessing algorithm which brightens the user's drawing while still retaining the details in the drawing. I deemed this necessary because I realized many of the drawings were not bright when converted to grayscale. I experimented by drawing the same image in magenta and in black. The magenta image was too dark when preprocessed and the network was unable to correctly classify it as a bag. Meanwhile, black image had much brighter activations when converted to grayscale, and the network classified it correctly with ~90% confidence. This shows that the color of the input drawing was having a great effect on the network's final prediction. The brightening step in the image preprocessing seems to have helped with this, although I have only brief observations and no concrete data.